pageinspect: Try to fix some bugs in previous commit.
Commit 08bf6e529587e1e9075d013d859af2649c32a511 seems not to have
used the correct *GetDatum and PG_GETARG_* macros for the SQL types
in some cases, and some of the SQL types seem to have been poorly
chosen, too.  Try to fix it.  I'm not sure if this is the reason
why the buildfarm is currently unhappy with this code, but it
seems like a good place to start.

Buildfarm unhappiness reported by Tom Lane.

+++ contrib/pageinspect/pageinspect--1.5--1.6.sql
@@ -19,14 +19,14 @@ LANGUAGE C STRICT PARALLEL SAFE;
 -- hash_page_stats()
 --
 CREATE FUNCTION hash_page_stats(IN page bytea,
-    OUT live_items smallint,
-    OUT dead_items smallint,
-    OUT page_size smallint,
-    OUT free_size smallint,
+    OUT live_items int4,
+    OUT dead_items int4,
+    OUT page_size int4,
+    OUT free_size int4,
     OUT hasho_prevblkno int8,
     OUT hasho_nextblkno int8,
     OUT hasho_bucket int8,
-	OUT hasho_flag smallint,
+	OUT hasho_flag int4,
 	OUT hasho_page_id int4)
 AS 'MODULE_PATHNAME', 'hash_page_stats'
 LANGUAGE C STRICT PARALLEL SAFE;
